Services y controllers

+++ smartnetwork-backend/src/main/java/com/smartnetwork/backend/Service/AddressService.java
@@ -1,7 +1,9 @@
 package com.smartnetwork.backend.Service;
 
 import com.smartnetwork.backend.Repository.AddressRepository;
+import com.smartnetwork.backend.Repository.DispositivoRepository;
 import com.smartnetwork.backend.domain.Entity.Address;
+import com.smartnetwork.backend.domain.Entity.Dispositivo;
 import org.springframework.stereotype.Service;
 import org.springframework.web.client.RestTemplate;
 
@@ -11,29 +13,88 @@ import java.util.Optional;
 @Service
 public class AddressService {
     private final AddressRepository addressRepository;
+    private final DispositivoRepository dispositivoRepo;
+    private final DispositivoService dispositivoService;
     private final RestTemplate restTemplate = new RestTemplate();
 
-    public AddressService(AddressRepository addressRepository) {
+    public AddressService(AddressRepository addressRepository, DispositivoRepository dispositivoRepo,DispositivoService dispositivoService) {
         this.addressRepository = addressRepository;
+        this.dispositivoRepo = dispositivoRepo;
+        this.dispositivoService = dispositivoService;
     }
 
-    public Address save(Address address){
+    public Address create(Address address, String username){
+        Dispositivo dispositivo = dispositivoRepo
+                .findById(address.getDispositivo().getId())
+                .orElseThrow(() -> new RuntimeException("Dispositivo no existe"));
+
+        // 🔐 Seguridad básica: el dispositivo es del usuario
+        if (!dispositivo.getUsuario().getUsername().equals(username)) {
+            throw new RuntimeException("No autorizado");
+        }
+
+        address.setDispositivo(dispositivo);
+
+        //FALTA COMPLETAR cearAddress en DispositivoService
+        try {
+           dispositivoService.crearAddress(dispositivo,address);
+        } catch (Exception e) {
+            throw new RuntimeException(e);
+        }
+
         return addressRepository.save(address);
+
     }
 
-    public List<Address> findAll(){
-        return addressRepository.findAll();
+    public List<Address> findAllByDispositivo(Long dispositivoId, String username) {
+
+        Dispositivo dispositivo = dispositivoRepo
+                .findById(dispositivoId)
+                .orElseThrow(() -> new RuntimeException("Dispositivo no existe"));
+
+        // 🔐 Seguridad: comprobar propietario
+        if (!dispositivo.getUsuario().getUsername().equals(username)) {
+            throw new RuntimeException("No autorizado");
+        }
+
+        return addressRepository.findByDispositivoId(dispositivoId);
     }
 
-    public void delete(Address address){
+    public void delete(Address address, String username, Long dispositivoId){
+        Dispositivo dispositivo = dispositivoRepo
+                .findById(dispositivoId)
+                .orElseThrow(() -> new RuntimeException("Dispositivo no existe"));
+
+        // 🔐 Seguridad: comprobar propietario
+        if (!dispositivo.getUsuario().getUsername().equals(username)) {
+            throw new RuntimeException("No autorizado");
+        }
         addressRepository.delete(address);
     }
 
-    public Optional<Address> findById(Long id){
-        return addressRepository.findById(id);
+    public Optional<Address> findById(String username, Long dispositivoId, Long addressId) {
+
+        Dispositivo dispositivo = dispositivoRepo
+                .findById(dispositivoId)
+                .orElseThrow(() -> new RuntimeException("Dispositivo no existe"));
+
+        // 🔐 Seguridad: comprobar propietario
+        if (!dispositivo.getUsuario().getUsername().equals(username)) {
+            throw new RuntimeException("No autorizado");
+        }
+
+        return addressRepository.findByIdAndDispositivoId(addressId, dispositivoId);
     }
 
-    public Address update(Address address){
+    public Address update(Address address, String username,Long dispositivoId){
+        Dispositivo dispositivo = dispositivoRepo
+                .findById(dispositivoId)
+                .orElseThrow(() -> new RuntimeException("Dispositivo no existe"));
+
+        // 🔐 Seguridad: comprobar propietario
+        if (!dispositivo.getUsuario().getUsername().equals(username)) {
+            throw new RuntimeException("No autorizado");
+        }
         return addressRepository.save(address);
     }
 }
